js/youtube: Remove $scope.apply wrapper for state change.
I am not sure if this is correct. The general rule if I am not mistaken
for appply is that it should be used for everything outside angular.
Well the state change is only triggered outside angular and the function
plus the called are angular objects that probably should be watched.

I ll test more and maybe I ll revert this