controllers/jukebox: Change tactic due to more buffer need.
I also need a callback here. Like when to start. In order to sync,
and give it a  5sec eg time back so when he requests the next it will be ready.
Of course this can be requested when he is manually seeking but then he will
get the current track at the correct (-5sec) time again.